Stop trusting the hand-written mirror of Pipedrive's types
The upstream types in internal/pipedrive are a hand-written mirror of
what Pipedrive returns, and nothing checked them. Holding them against
Pipedrive's own v2 OpenAPI description found three defects that had
shipped, each confirmed against a live workspace:

- Pipeline read `active` and Stage read `active_flag`. v2 returns
  neither — both carry is_deleted — so both decoded to false for every
  record and list_pipelines told the model every pipeline in the
  workspace was inactive. Both spellings are v1 vocabulary that came
  across in the port, under a comment claiming they had been confirmed
  against the live v2 endpoints.
- Organization declared people_count without the include_fields request
  that makes v2 send it, so it was absent from every row while three
  tool descriptions promised it.
- Deal.Probability was *int where the spec says `number`. Go refuses
  12.5 into an int and client.go returns that error, so one such deal
  failed the whole page rather than one row.

spec_test.go now walks the json tags against a fixture derived from the
spec. It is a test, not a generator: the spec has zero $ref, so a
generator emits an anonymous struct per endpoint rather than one Deal,
and models-only emits no response types at all. That measurement is what
closed the hand-rolled-versus-generated spike, recorded in CONTRIBUTING
and docs/architecture.md.

Two exemptions are recorded with reasons, both verified live —
is_writable, which the *Fields endpoints return and the write guard
depends on while the spec has never declared it, and people_count.

BREAKING CHANGE: probability is a number rather than an integer, on the
deal summary and on manage_deal's input. Pipedrive declares it that way
and a fractional value previously failed the read outright.

+++ internal/pipedrive/deals.go
@@ -36,16 +36,16 @@ type ListDealsOptions struct {
 // pipeline, ...). The tool layer enforces title non-empty client-side
 // so a typo surfaces as [validation] instead of an upstream 400.
 type CreateDealRequest struct {
-	Title             string  `json:"title"`
-	Value             float64 `json:"value,omitempty"`
-	Currency          string  `json:"currency,omitempty"`
-	PipelineID        int64   `json:"pipeline_id,omitempty"`
-	StageID           int64   `json:"stage_id,omitempty"`
-	OwnerID           int64   `json:"owner_id,omitempty"`
-	PersonID          int64   `json:"person_id,omitempty"`
-	OrgID             int64   `json:"org_id,omitempty"`
-	ExpectedCloseDate string  `json:"expected_close_date,omitempty"` // YYYY-MM-DD
-	Probability       *int    `json:"probability,omitempty"`         // 0-100; nil = use stage default
+	Title             string   `json:"title"`
+	Value             float64  `json:"value,omitempty"`
+	Currency          string   `json:"currency,omitempty"`
+	PipelineID        int64    `json:"pipeline_id,omitempty"`
+	StageID           int64    `json:"stage_id,omitempty"`
+	OwnerID           int64    `json:"owner_id,omitempty"`
+	PersonID          int64    `json:"person_id,omitempty"`
+	OrgID             int64    `json:"org_id,omitempty"`
+	ExpectedCloseDate string   `json:"expected_close_date,omitempty"` // YYYY-MM-DD
+	Probability       *float64 `json:"probability,omitempty"`         // 0-100; nil = use stage default
 
 	// CustomFields is keyed by the 40-char hash Pipedrive stores,
 	// with dropdown values as option ids: the shape FieldCache.Encode
@@ -86,7 +86,7 @@ type UpdateDealRequest struct {
 	PersonID          *int64   `json:"person_id,omitempty"`
 	OrgID             *int64   `json:"org_id,omitempty"`
 	ExpectedCloseDate *string  `json:"expected_close_date,omitempty"`
-	Probability       *int     `json:"probability,omitempty"`
+	Probability       *float64 `json:"probability,omitempty"`
 	Status            *string  `json:"status,omitempty"`
 	LostReason        *string  `json:"lost_reason,omitempty"`
